Changed the global constants for the window size to floats. Removed the initial positioning of the paddles and ball.
I figured I did alot of casting the two globals into floats, so why not just change them to floats to save on casting them. Only have to cast them once now into unsigned ints for the sf::Videomode. Then I also took away the power of the paddles and ball to position themselves and gave that to the constructor of the game class.